DomDecOT.AbstractPlan — Type

AbstractPlan

Super type of all implementations of OT plans.

DomDecOT.DomDecPlan — Type

DomDecPlan(mu::AbstractMeasure, nu::AbstractMeasure, gamma,
            cellsize::Int[, basic_cells::Vector, 
            composite_cells::Vector, partitions::Vector,
            alphas::Vector, betas::Vector, 
            epsilon::Float64, partk::Int)

A DomDecPlan is a struct that keeps track of the status of the domain decomposition algorithm in an effient manner. Its arguments are

mu: AbstractMeasure representing the X marginal

nu: AbstractMeasure representing the Y marginal

gamma: gamma[i] is a sparse vector representing the current marginal of basic cell i. Alternatively, gamma can also be a sparse matrix representing the full initial plan.

cellsize::Int: maximum size of the basic cells (along all dimensions).

basic_cells: basic_cells[i] is the indices of the atoms in mu that are merged together to form a basic cell.

composite_cells: composite_cells[k][j] refers to the group of basic cells that constitute the j-th subdomain of the k-th partition.

partitions: partitions[k][j] are the indices of all the X atoms that constitute the first marginal during j-th subdomain of partition k. It equals vcat([basic_cells[composite_cells[k][j]]...)

alphas: X-dual potential on each subdomain. alphas[k][j] has the same length as partitions[k][j].

betas: Y-dual potentials on each subdomain.

epsilon: last global epsilon used to solve the cell problems.

partk: index of the last partition whose subdomains were solved.

DomDecOT.PD_gap — Method

PD_gap(P, c,  ε)

Primal-dual gap of plan P for the cost c and regularization ε.

DomDecOT.balance! — Function

balance!(Q, μ, threshold, force_balancing = true)

Apply balance! with threshold threshold on pairs of columns of Q until each column Q[:,i] has mass μ. When force_balancing == true, if a first pass wasn't succesful, a second pass is attempted seeting threshold = 0.

DomDecOT.balance! — Method

balance!(a, b, δ, threshold; force_balancing = false)

Transfer δ mass from vector a to b. The recipient entries must be larger than threshold; if it is not achieved a warning is thrown.

DomDecOT.cell_PD_gap — Method

cell_PD_gap(P, k, j, c, ε)

Primal-dual gap in cell j of partition k.

DomDecOT.domdec_logsinkhorn! — Method

domdec_logsinkhorn!(β, α, νJ, νI, μJ, C, ε; kwargs...)

Solve cell problem using the log-domain Sinkhorn algorithm of the MultiScaleOT library.

Arguments

β: initial Y dual potential

α: initial X dual potential

νJ: Y cell marginal

νI: global Y marginal supported on the same points as νJ_global

μJ: X cell marginal

C: cost matrix. It is transformed inplace to yield the primal plan

ε: regularization

DomDecOT.domdec_sinkhorn_autofix_log! — Method

domdec_sinkhorn_autofix_log!(β, α, νJ, νI, μJ, C, ε; kwargs...)

Attempt to solve the cell problem by calling sinkhorn_stabilized!. If the algorithm errors, fall back to logsinkhorn!.

Arguments

β: initial Y dual potential

α: initial X dual potential

νJ: Y cell marginal

νI: global Y marginal supported on the same points as νJ_global

μJ: X cell marginal

C: cost matrix. It is transformed inplace to yield the primal plan

ε: regularization

DomDecOT.domdec_sinkhorn_stabilized! — Method

domdec_sinkhorn_stabilized!(β, α, νJ, νI, μJ, C, ε; kwargs...)

Solve cell problem using the stabilized Sinkhorn algorithm of the MultiScaleOT library.

Arguments

β: initial Y dual potential

α: initial X dual potential

νJ: Y cell marginal

νI: global Y marginal supported on the same points as νJ_global

μJ: X cell marginal

C: cost matrix. It is transformed inplace to yield the primal plan

ε: regularization

DomDecOT.dual_score — Method

dual_score(P, c,  ε)

Compute the dual score of P for the entropic problem with cost function c and regularization ε, using only the non-zero entries of P. For the dense dual score, consider converting first P to a sparse matrix and then using the routines of MultiScaleOT.

DomDecOT.fix_beta! — Method

fix_beta!(β, α, C, νJ, νI, μJ, ε)

Set β to the conjugate of α, taking into account the cell parameters. Stable implementation.

DomDecOT.fix_beta! — Method

fix_beta!(β, N::Int)

Check if the Y-dual parameter has the appropriate size. If not, turns it into a zeros(N). Return true if β was of the appropriate length, false otherwise

DomDecOT.get_alpha_field — Method

get_alpha_field(P, k)

Glue cell duals of partition k to form a global dual.

DomDecOT.get_alpha_graph — Method

get_alpha_graph(P::DomDecPlan, alpha_diff, cellsize)

Compute Helmholtz decomposition on, alpha_diff, averaging on basic cells, as explained in https://arxiv.org/abs/2001.10986, Section 6.3.

DomDecOT.get_basic_and_composite_cells — Method

get_basic_and_composite_cells(gridshape, cellsize)

Get the basic and composite cells corresponding to partitioning a grid of shape gridshape into basic cells of size at most cellsize × ... × cellsize, and then joining groups of 2^D adjacent basic cells together to form composite cells compA. compB is obtained in an analogous manner, but with an offset of 1 along each axes.

Examples

DomDecOT.get_cell_Y_marginal — Method

get_cell_Y_marginal(P, J)

Compute the total Y-marginal of a set of basic cells J

DomDecOT.get_cost_matrix — Method

get_cost_matrix(P, c, J, I)

Return the cost matrix corresponding to points Y[:,I] and X[:,J].

DomDecOT.get_discrete_gradient — Method

get_discrete_gradient(nx, ny, nz)

Get (transpose of) X-, Y- and Z-gradient matrices for a discrete graph of size (nx, ny, nz).

DomDecOT.get_discrete_gradient — Method

get_discrete_gradient(nx, ny)

Get (transpose of) X- and Y-gradient matrices for a discrete graph of size (nx, ny).

Relation with the Python library

If we define in Julia

gx, gy = get_discrete_gradient(nx, ny)

and in Python

GX, GY = Common.getDiscreteGradients(ny, nx)

Then GX, GY == gx', gy'

DomDecOT.get_discrete_gradient — Method

get_discrete_gradient(nx)

Get (transpose of) X-gradient matrices for a one dimensional lattice of size nx.

DomDecOT.get_pairwise_delta — Method

get_pairwise_delta(d1, d2)

For a pair of values d1, d2 representing the offset of mass of two vectors, computes the maximum possible transfer of mass that does not make any vector worse than what it is currently. For example, if both vectors have mass in excess or in defect, it returns 0; if one has in excess and another in defect, it returns the smallest of these adjustments, with the corresponding sign.

Examples

julia> getpairwisedelta(1,2), getpairwisedelta(-1, -3) (0, 0)

julia> getpairwisedelta(1,-2) 1

julia> getpairwisedelta(-1,2) -1

DomDecOT.get_partition — Method

get_partition(basic_cells, composite_cells)

Compute a partition by joining the basic cells corresponding to each composite cells.

Examples

julia> DD.get_partition([[1], [2, 3], [4]], [[1, 2], [3]])
2-element Vector{Vector{Int64}}:
 [1, 2, 3]
 [4]

DomDecOT.iterate! — Method

iterate!(P, c, solver, params::NamedTuple)

Perform params[:iters] inplace iterations of the domain decomposition algorithm on the plan P, using the cost c and the inner solver solver. If params[:parallel_iteration] is true, runs the iterations in parallel.

DomDecOT.iterate_parallel! — Method

iterate_parallel!(P, k, c, solver, params)

Run an (inplace), parallel half iteration k of domain decomposition on plan P, with cost c, solver solver and ceratain params.

DomDecOT.iterate_serial! — Method

iterate_serial!(P, k, c, solver, params)

Run an (inplace), serial half iteration k of domain decomposition on plan P, with cost c, solver solver and ceratain params.

DomDecOT.plan_to_dense_matrix — Function

plan_to_dense_matrix(P, c[, k, balancing = true])

Turn P into a dense matrix using the dual potentials of partition k.

DomDecOT.plan_to_sparse_matrix — Function

plan_to_sparse_matrix(P, c[, k, balancing = true])

Turn P into a sparse matrix using the dual in the last iteration.

DomDecOT.primal_score — Method

primal_score(P, c, ε)

Compute the primal score of P for the entropic problem with cost function c and regularization ε, using only the non-zero entries of P.

DomDecOT.reduce_cellplan_to_cellmarginals — Method

reduce_cellplan_to_cellmarginals(P::DomDecPlan, PJ::AbstractMatrix, k, j, μJ)

Take the matrix PJ, that is assumed to be the current plan on cell k of partition k and extract its cell marginals.

DomDecOT.reduce_to_cells — Method

reduce_to_cellsize(γ0::AbstractMatrix, basic_cells)

Add together marginals corresponding to the same basic cell.

DomDecOT.reduce_to_cells — Method

reduce_to_cellsize(γ0::AbstractVector, basic_cells)

Add together marginals corresponding to the same basic cell.

DomDecOT.reduce_to_cells — Method

reduce_to_cells(gamma, gridshape, cellSize)

Add together marginals corresponding to the same basic cell, for given shapeX and cellSize.

DomDecOT.refine_plan — Method

refine_plan(P::DomDecPlan, 
            μH::MultiScaleMeasure{GridMeasure{D}}, 
            νH::MultiScaleMeasure, 
            k::Int;
            consistency_check = true) where D

Refine the plan P to match the refinement of the X marginal μH[k-1] -> μH[k] and of the Y marginal νH[k-1] -> νH[k]

DomDecOT.refine_submeasure — Method

refine_submeasure(v, u, u2, Js)

Refine the measure v with respect to the Y-cells given in Js by matching the pattern induced by the refinement u -> u2.

DomDecOT.smooth_alpha_and_beta_fields — Method

smooth_alpha_and_beta_field(P::DomDecPlan, c)

Compute a smooth version of the duals of P by performing a Helmholtz decomposition (see https://arxiv.org/abs/2001.10986, Section 6.3 for details)

DomDecOT.smooth_alpha_field — Method

smooth_alpha_field(P::DomDecPlan, cellsize)

Compute a smooth version of the X dual of P by performing a Helmholtz decomposition (see https://arxiv.org/abs/2001.10986, Section 6.3 for details)

DomDecOT.solvecell! — Method

solvecell!(P::DomDecPlan, k, j, c, solver, params)

Perform an (inplace) cell iteration on cell P.partitions[k][j] of the plan P, using the inner solver and the given params.

DomDecOT.truncate! — Method

truncate!(A, [μJ, ] threshold)

Drop values of ν smaller or equal than threshold. Acts inplace. If μJ is given and A is a matrix, call

truncate!(A[:,i], μJ[i]*threshold)

for each column of A.

DomDecOT.truncate! — Method

truncate!(P::AbstractPlan, rel_threshold)

Drop values of the plan P that are smaller than rel_threshold with respect to the cell mass. Acts inplace.

DomDecOT.update_cell_plan! — Method

update_cell_plan!(P::DomDecPlan, PJ, I, k, j, μJ[;
        balance = true, truncate = true,
        truncate_Ythresh = 1e-16, truncate_Ythresh_rel=true])

Update cell j of partition k with the given cell plan PJ.

DomDecOT.update_cell_plan! — Method

update_cell_plan!(P::DomDecPlan, P_basic, k, j, I)

Update cell j of partition k of the plan P using the columns in P_basic as the basic cell marginals. I is the real support of the columns of P_basic

DomDecOT.view_X — Method

view_X(P, J)

Return a view of the X points in subset J of P.mu.points.

DomDecOT.view_Y — Method

view_cell_Y(P, I)

Return a view of the Y points with index I.