Library

Plan

The current state of the domain decomposition algorithm is stored in an AbstractPlan. The only AbstractPlan provided is called DomDecPlan, but new ones can be build if needed. Though it is not strictly necessary to understand its internals to use it, we leave it here for reference:

Missing docstring.

Missing docstring for DomDecPlan. Check Documenter's build log for details.

It is important to note that a DomDecPlan does not store the coupling explicitly. Instead, it stores the marginals of the basic cells of the coupling, from which one can recover the Y-marginals of the partitions where the subproblems are solved. This, together with the dual marginals (that are stored in the fields alphas and betas) allows to reconstruct the whole coupling if needed, while only using a fraction of the total memory.

Iteration

DomDecOT.iterate! — Function

iterate!(P, c, solver, params::NamedTuple)

Perform params[:iters] inplace iterations of the domain decomposition algorithm on the plan P, using the cost c and the inner solver solver. If params[:parallel_iteration] is true, runs the iterations in parallel.

Here it is to note that any cost function c(x,y) consistent with the columns of P.mu.points and P.nu.points will work.

Note

A note about parameters

Parameters are to be packaged in a NamedTuple. This is a type-stable analog of a dictionary that can be build as follows

julia> (; epsilon = 1.0, truncation_thres = 1e-14, verbose = true)
(epsilon = 1.0, truncation_thres = 1e-14, verbose = true)

DomDecOT.jl provides easy ways to generate and operate with this tuples, that will be covered in section [ref]

Solvers

The subproblems are solved with one of the following solvers, that build upon those provided by MultiScaleOT.jl.

DomDecOT.domdec_sinkhorn_stabilized! — Function

domdec_sinkhorn_stabilized!(β, α, νJ, νI, μJ, C, ε; kwargs...)

Solve cell problem using the stabilized Sinkhorn algorithm of the MultiScaleOT library.

Arguments

β: initial Y dual potential

α: initial X dual potential

νJ: Y cell marginal

νI: global Y marginal supported on the same points as νJ_global

μJ: X cell marginal

C: cost matrix. It is transformed inplace to yield the primal plan

ε: regularization

DomDecOT.domdec_logsinkhorn! — Function

domdec_logsinkhorn!(β, α, νJ, νI, μJ, C, ε; kwargs...)

Solve cell problem using the log-domain Sinkhorn algorithm of the MultiScaleOT library.

Arguments

β: initial Y dual potential

α: initial X dual potential

νJ: Y cell marginal

νI: global Y marginal supported on the same points as νJ_global

μJ: X cell marginal

C: cost matrix. It is transformed inplace to yield the primal plan

ε: regularization

DomDecOT.domdec_sinkhorn_autofix_log! — Function

domdec_sinkhorn_autofix_log!(β, α, νJ, νI, μJ, C, ε; kwargs...)

Attempt to solve the cell problem by calling sinkhorn_stabilized!. If the algorithm errors, fall back to logsinkhorn!.

Arguments

β: initial Y dual potential

α: initial X dual potential

νJ: Y cell marginal

νI: global Y marginal supported on the same points as νJ_global

μJ: X cell marginal

C: cost matrix. It is transformed inplace to yield the primal plan

ε: regularization

Obtaining the coupling and duals

One we have performed a series of iterations, we are usually interested in obtaining the actual coupling.

DomDecOT.plan_to_dense_matrix — Function

plan_to_dense_matrix(P, c[, balancing = true])

Turn P into a dense matrix using the dual potentials of the last iteration.

DomDecOT.plan_to_sparse_matrix — Function

plan_to_sparse_matrix(P, c[, k, balancing = true])

Turn P into a sparse matrix using the dual in the last iteration.

As well as the primal coupling, we are usually interested in obtaining the duals. Note that in a DomDecPlan only the duals of each partition are saved, so one must glue them all together to obtain global potentials. This is achieved with the following function.

DomDecOT.smooth_alpha_and_beta_fields — Function

smooth_alpha_and_beta_field(P::DomDecPlan, c)

Compute a smooth version of the duals of P by performing a Helmholtz decomposition (see https://arxiv.org/abs/2001.10986, Section 6.3 for details)

Computing scores

One can compute the primal and dual scores directly on the DomDecPlan by calling the following functions.

DomDecOT.primal_score — Function

primal_score(P, c, ε)

Compute the primal score of P for the entropic problem with cost function c and regularization ε, using only the non-zero entries of P.

DomDecOT.dual_score — Function

dual_score(P, c,  ε)

Compute the dual score of P for the entropic problem with cost function c and regularization ε, using only the non-zero entries of P. For the dense dual score, consider converting first P to a sparse matrix and then using the routines of MultiScaleOT.

Missing docstring.

Missing docstring for PD_gap. Check Documenter's build log for details.

Alternatively, if the global primal and duals are already available in matrix form, one can also use the functions primal_score_sparse and dual_score_sparse of the MultiScaleOT.jl library.

Refinement

Domain decomposition becomes really powerful when combined with a hierarchical scheme. For this it is necessary to, from a coarse solution to layer k-1, provide a feasible initialization for the finer layer k. This can be achieved with refine_plan, which also initializes the duals of the new plan by interpolating between those in the old one.

DomDecOT.refine_plan — Function

refine_plan(P::DomDecPlan, 
            μH::MultiScaleMeasure{GridMeasure{D}}, 
            νH::MultiScaleMeasure, 
            k::Int;
            consistency_check = true) where D

Refine the plan P to match the refinement of the X marginal μH[k-1] -> μH[k] and of the Y marginal νH[k-1] -> νH[k]